backend/requirements.txt
# Web Framework
flask>=2.3
flask-cors>=4.0

# Face Recognition (FaceNet - PyTorch)
torch>=2.0,<3.0
torchvision>=0.15
facenet-pytorch>=2.5

# Image Processing
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0

# Encryption
pycryptodome>=3.19

# Liveness Detection (facial landmarks)
dlib-bin>=19.24

# Configuration
python-dotenv>=1.0

# Database
# SQLite is built-in